moved the question tracking to the question component
