+++ client/src/components/Question/index.js
@@ -1,9 +1,37 @@
-import React from "react";
+import React, {useState } from "react";
 import "./style.css"
-
+import { useDispatch, useSelector } from "react-redux";
+import { RECORD_ANSWER } from "../../utils/actions";
 
 /// State to hold user input
-function Question({ question, action }) {
+function Question() {
+  const questions = useSelector(state => state.workoutQuestions);
+  if (!questions) {
+    window.location.assign('/');
+  }
+  const dispatch = useDispatch();
+  // if token is valid it will return user data
+  let [questionIdIndex, setQuestionIdIndex] = useState(0);
+  let [questionIds, setQuestionIds] = useState(Object.keys(questions));
+  const question=questions[questionIds[questionIdIndex]];
+  const recordAnswer = (event) => {
+    event.preventDefault();
+    dispatch({
+      type: RECORD_ANSWER,
+      answer: { questionId: questionIds[questionIdIndex], answer: document.getElementById(questionIds[questionIdIndex]).value }
+    })
+    if (event.target.id === 'prev' && questionIdIndex > 0) {
+      setQuestionIdIndex(questionIdIndex - 1);
+    }
+    if (event.target.id === 'next') {
+      if (questionIdIndex < questionIds.length - 1) {
+        setQuestionIdIndex(questionIdIndex + 1);
+      } else {
+        alert("BuildWorkout")
+      }
+    }
+  }
+
   let input = ""
   switch (question.fieldType) {
     case "text":
@@ -31,8 +59,8 @@ function Question({ question, action }) {
         <form style={{paddingTop: '100px'}}>
           question: {question.question}<br></br>
           {input}<br></br>
-          <button id="prev" onClick={action}>Previous</button>
-          <button id="next" onClick={action}>Next</button>
+          <button id="prev" onClick={recordAnswer}>Previous</button>
+          <button id="next" onClick={recordAnswer}>Next</button>
         </form>
       </div>
     </>
